Refactor getting function expression node path
