+++ lib/collections/ArrowFunctionExpression.js
@@ -7,7 +7,7 @@ const Node = require("./Node");
 
 const types = recast.types.namedTypes;
 
-const ArrowFunctionExpression = types.namedTypes.ArrowFunctionExpression;
+const ArrowFunctionExpression = types.ArrowFunctionExpression;
 
 /**
 * @mixin
