+++ lib/collections/LegacyAction.js
@@ -6,13 +6,17 @@ const { fromPaths } = require("jscodeshift/dist/Collection");
 
 const types = recast.types.namedTypes;
 
-const Node = recast.types.namedTypes.Node;
+const File = types.File;
 
 
 /**
 * @mixin
 */
 const traversalMethods = {
+  getFunctionExpression: function() {
+    return fromPaths([this.get("program", "body", 0, "expression")], this);
+  },
+
   /**
    * Finds occurences of `auths` member expressions
    * 
@@ -21,7 +25,7 @@ const traversalMethods = {
    * @return {Collection}
    */
   getAuthMemberExpressions: function() {
-    return fromPaths([this.get("program", "body", 0, "expression")], this)
+    return this.getFunctionExpression()
       .getVariableInstances(1)
       .map((path) => path.parent)
       .filter((path) => types.MemberExpression.check(path.value));
@@ -35,7 +39,7 @@ const traversalMethods = {
    * @return {Collection}
    */
   getAuthVariableDeclarators: function() {
-    return fromPaths([this.get("program", "body", 0, "expression")], this)
+    return this.getFunctionExpression()
       .getVariableInstances(1)
       .map((path) => path.parent)
       .filter((path) => types.VariableDeclarator.check(path.value))
@@ -94,7 +98,7 @@ const traversalMethods = {
 
 function register() {
   registerArrowFunctionExpression();
-  Collection.registerMethods(traversalMethods, Node);
+  Collection.registerMethods(traversalMethods, File);
 }
 
 exports.register = once(register);
