+++ lib/transforms/auths-to-auth.js
@@ -1,7 +1,7 @@
 const { register } = require("../collections/LegacyAction");
 register();
 
-// replace `auths.github` with `this.github.$auth`
+// replace `auths.__a` with `this.__a.$auth`
 // replace `const { __a: __b } = auths` with `const __b = this.__a.$auth`
 
 module.exports = function(fileInfo, api, options) {
