+++ lib/transforms/checkpoint-to-db-get.js
@@ -11,14 +11,14 @@ module.exports = function(fileInfo, api, options) {
   ast
     .getCheckpointExpressions()
     .replaceWith(() => {
-      // this.db.get("$checkpoint");
+      // `this.db.get("$checkpoint")`
       return j.callExpression(
         j.memberExpression(
-          j.memberExpression(j.thisExpression(), j.identifier("db")), // this.db
-          j.identifier("get") // .get
-        ), [ // (
-          j.literal("$checkpoint"),  // "$checkpoint",
-        ] // )
+          j.memberExpression(j.thisExpression(), j.identifier("db")), // `this.db`
+          j.identifier("get") // `.get`
+        ), [ // `(`
+          j.literal("$checkpoint"),  // `"$checkpoint",`
+        ] // `)`
       );
     });
