+++ lib/transforms/checkpoint-to-db-set.js
@@ -11,15 +11,15 @@ module.exports = function(fileInfo, api, options) {
     .getCheckpointAssignmentExpressions()
     .replaceWith((path) => {
       const value = path.value.right;
-      // this.db.set("$checkpoint", value);
+      // `this.db.set("$checkpoint", value)`
       return j.callExpression(
         j.memberExpression(
-          j.memberExpression(j.thisExpression(), j.identifier("db")), // this.db
-          j.identifier("set") // .set
-        ), [ // (
-          j.literal("$checkpoint"),  // "$checkpoint",
-          value, // value
-        ] // )
+          j.memberExpression(j.thisExpression(), j.identifier("db")), // `this.db`
+          j.identifier("set") // `.set`
+        ), [ // `(`
+          j.literal("$checkpoint"),  // `"$checkpoint",`
+          value, // `value`
+        ] // `)`
       );
     });
