+++ lib/transforms/named-export-to-dollar-export.js
@@ -1,8 +1,7 @@
-const { fromPaths } = require("jscodeshift/dist/Collection");
 const { register } = require("../collections/ArrowFunctionExpression");
 register();
 
-// convert `this.__a = __b` to `$.export(__a, __b)` when this is bound to root
+// convert `this.__a = __b` to `$.export(__a, __b)`
 
 module.exports = function(fileInfo, api, options) {
   const j = api.jscodeshift;
