+++ lib/transforms/params-to-props.js
@@ -1,14 +1,11 @@
-const { fromPaths } = require("jscodeshift/dist/Collection");
-const { register } = require("../collections/ArrowFunctionExpression");
+const { register } = require("../collections/LegacyAction");
 register();
 
 module.exports = function(fileInfo, api, options) {
   const j = api.jscodeshift;
 
-  const ast = j(fileInfo.source);
-  const functionExpr = fromPaths([ast.get("program", "body", 0, "expression")], ast);
-
-  return functionExpr
+  return j(fileInfo.source)
+    .getFunctionExpression()
     .renameTo(0, "this")
     .toSource();
 };
