+++ lib/transforms/program-to-expression.js
@@ -1,7 +1,9 @@
+const { register } = require("../collections/LegacyAction");
+register();
+
 module.exports = function(fileInfo, api, options) {
   const j = api.jscodeshift;
 
   const ast = j(fileInfo.source);
-  return j(ast.get("program", "body", 0, "expression", "body"))
-    .toSource();
+  return j(ast.getFunctionExpression().get("body")).toSource();
 };
