+++ lib/transforms/send-to-dot-send.js
@@ -1,16 +1,13 @@
-const { fromPaths } = require("jscodeshift/dist/Collection");
-const { register } = require("../collections/Node");
+const { register } = require("../collections/LegacyAction");
 register();
 
-// convert $send to $.send
+// convert `$send` to `$.send`
 
 module.exports = function(fileInfo, api, options) {
   const j = api.jscodeshift;
 
-  const ast = j(fileInfo.source);
-  const functionExpr = fromPaths([ast.get("program", "body", 0, "expression")], ast);
-
-  functionExpr
+  const ast = j(fileInfo.source)
+    .getFunctionExpression()
     .getVariableOccurences("$send")
     .replaceWith(() => {
       return j.memberExpression(j.identifier("$"), j.identifier("send"));
